FAC-1231: Move `SpaceStatus` to common module and update references across services and tests
