+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/application/domain/space/SpaceStatus.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.users.application.domain;
+package org.flickit.assessment.common.application.domain.space;
 
 public enum SpaceStatus {
