+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/user/CreateUserService.java
@@ -2,7 +2,7 @@ package org.flickit.assessment.users.application.service.user;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.application.domain.space.SpaceType;
-import org.flickit.assessment.users.application.domain.SpaceStatus;
+import org.flickit.assessment.common.application.domain.space.SpaceStatus;
 import org.flickit.assessment.users.application.domain.SpaceUserAccess;
 import org.flickit.assessment.users.application.port.in.user.CreateUserUseCase;
 import org.flickit.assessment.users.application.port.out.space.CreateSpacePort;
